🖥️  17: Chronospatial Computer
Part 1 was more or less easy, but part 2 was heavy.

The solution now is actually the solution from
https://github.com/zebalu/advent-of-code-2024/blob/master/solution/src/main/java/io/github/zebalu/aoc2024/Day17.java

While adapting the solution, I found a critical bug in my division
implementation. I'd read and adapted "truncated to integer" into the
Java world literally, but that was a mistake. The should be not
casting. Although this flaw influences working with high numbers
(hello part2), my own solutions still seems to be lacking a bit
(pun intended).