📋 Day 19: Linen Layout
The part1 was easy, but part2 is not suitable for recursion/iteration of all combinations.

A cache was obvious, but won't work at the end. Description of this solution
https://github.com/jwtrhs/advent-of-code/blob/main/2024/19/main.py was exactly my idea
(literally), but I did not think to sum up the findings in the count method directly.
Also I tried always the opposite way (first the longer strings).  Spend some time for
finding long-running code with doubled-matchings…  😩