⌨️ Day 21: Keypad Conundrum
After three different strategies which all have one in common having *the same
wrong output for the sample* 😩, I look up for some hints. https://aoc.csokavar.hu/2024/21/

First of all, because of the nature each step must be encoded (which I had already),
it is unnecessary to look for diagonal moves because they will be to expensive after
all. That makes the options far easier.
Secondly, the complexity can be reduced not storing state in the device
(I had the current position stored there), but simply use the state in the
encoding steps.
And at last: do not return the actual movements, but reduce to the actual costs. That improves
the memory costs dramatically.

As I already had the guess the number of layers/robots will be changed in part2,
my preps for this were not wrong, but not enough. Adding a cache is the way.